InputColumnsWithDataType
