+++ src/query/expression/src/aggregate/aggregate_function.rs
@@ -14,9 +14,6 @@
 
 use std::alloc::Layout;
 use std::fmt;
-use std::ops::Index;
-use std::ops::Range;
-use std::slice::SliceIndex;
 use std::sync::Arc;
 
 use databend_common_arrow::arrow::bitmap::Bitmap;
@@ -27,7 +24,7 @@ use crate::types::binary::BinaryColumnBuilder;
 use crate::types::DataType;
 use crate::Column;
 use crate::ColumnBuilder;
-use crate::DataBlock;
+use crate::InputColumns;
 use crate::Scalar;
 
 pub type AggregateFunctionRef = Arc<dyn AggregateFunction>;
@@ -169,121 +166,4 @@ pub trait AggregateFunction: fmt::Display + Sync + Send {
     fn convert_const_to_full(&self) -> bool {
         true
     }
-}
-
-#[derive(Copy, Clone)]
-pub enum InputColumns<'a> {
-    Slice(&'a [Column]),
-    Block(BlockProxy<'a>),
-}
-
-impl Index<usize> for InputColumns<'_> {
-    type Output = Column;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        match self {
-            Self::Slice(slice) => slice.index(index),
-            Self::Block(BlockProxy { args, data }) => {
-                data.get_by_offset(args[index]).value.as_column().unwrap()
-            }
-        }
-    }
-}
-
-impl<'a> InputColumns<'a> {
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    pub fn len(&self) -> usize {
-        match self {
-            Self::Slice(s) => s.len(),
-            Self::Block(BlockProxy { args, .. }) => args.len(),
-        }
-    }
-
-    pub fn slice<I>(&self, index: I) -> InputColumns<'_>
-    where I: SliceIndex<[usize], Output = [usize]> + SliceIndex<[Column], Output = [Column]> {
-        match self {
-            Self::Slice(s) => Self::Slice(&s[index]),
-            Self::Block(BlockProxy { args, data }) => Self::Block(BlockProxy {
-                args: &args[index],
-                data,
-            }),
-        }
-    }
-
-    pub fn iter(&self) -> InputColumnsIter {
-        match self {
-            Self::Slice(s) => InputColumnsIter {
-                iter: 0..s.len(),
-                this: self,
-            },
-            Self::Block(BlockProxy { args, .. }) => InputColumnsIter {
-                iter: 0..args.len(),
-                this: self,
-            },
-        }
-    }
-
-    pub fn new_block_proxy(args: &'a [usize], data: &'a DataBlock) -> InputColumns<'a> {
-        Self::Block(BlockProxy { args, data })
-    }
-}
-
-pub struct InputColumnsIter<'a> {
-    iter: Range<usize>,
-    this: &'a InputColumns<'a>,
-}
-
-unsafe impl<'a> std::iter::TrustedLen for InputColumnsIter<'a> {}
-
-impl<'a> Iterator for InputColumnsIter<'a> {
-    type Item = &'a Column;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.iter.next().map(|index| self.this.index(index))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.iter.size_hint()
-    }
-
-    fn count(self) -> usize
-    where Self: Sized {
-        self.iter.count()
-    }
-
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.iter.nth(n).map(|index| self.this.index(index))
-    }
-
-    fn last(self) -> Option<Self::Item>
-    where Self: Sized {
-        self.iter.last().map(|index| self.this.index(index))
-    }
-}
-
-impl<'a> From<&'a [Column]> for InputColumns<'a> {
-    fn from(value: &'a [Column]) -> Self {
-        InputColumns::Slice(value)
-    }
-}
-
-impl<'a, const N: usize> From<&'a [Column; N]> for InputColumns<'a> {
-    fn from(value: &'a [Column; N]) -> Self {
-        InputColumns::Slice(value.as_slice())
-    }
-}
-
-impl<'a> From<&'a Vec<Column>> for InputColumns<'a> {
-    fn from(value: &'a Vec<Column>) -> Self {
-        InputColumns::Slice(value)
-    }
-}
-
-#[derive(Copy, Clone)]
-pub struct BlockProxy<'a> {
-    args: &'a [usize],
-    data: &'a DataBlock,
 }
