+++ src/query/expression/src/lib.rs
@@ -51,6 +51,7 @@ mod evaluator;
 mod expression;
 pub mod filter;
 mod function;
+mod input_columns;
 mod kernels;
 mod property;
 mod register;
@@ -69,6 +70,7 @@ pub use crate::evaluator::*;
 pub use crate::expression::*;
 pub use crate::filter::*;
 pub use crate::function::*;
+pub use crate::input_columns::*;
 pub use crate::kernels::*;
 pub use crate::property::*;
 pub use crate::register::*;
