+++ src/query/expression/tests/it/input_columns.rs
